Rename `evo.objects.loader` -> `evo.objects.parquet`
