+++ packages/evo-objects/src/evo/objects/parquet/__init__.py
@@ -14,7 +14,7 @@ try:
 except ImportError:
     raise ImportError("The 'pyarrow' package is required to use ParquetLoader") from None
 
-from .parquet_loader import ParquetLoader
+from .loader import ParquetLoader
 from .types import ArrayTableInfo, LookupTableInfo, TableInfo
 
 __all__ = [
